时间工具类，实现metaObjectHandler，插入数据时填充create_tiem，修改数据时填充modify_time

+++ muxige-core/src/main/java/org/minpoint/muxige/core/pojo/DateEntity.java
@@ -20,10 +20,10 @@ public class DateEntity extends BaseEntity{
      * 创建时间
      */
     @TableField( value = "create_time", fill = FieldFill.INSERT)
-    private Integer createTime;
+    private Long createTime;
     /**
      * 修改时间
      */
     @TableField( value = "modify_time", fill = FieldFill.UPDATE)
-    private Integer modifyTime;
+    private Long modifyTime;
 }
